+++ sdk/node/src/__tests__/selectbox.test.ts
@@ -1,77 +1,87 @@
 import { expect, test } from 'vitest';
 import { v4 as uuidv4 } from 'uuid';
-import { RadioState } from './internal/session/state/radio';
-import { convertRadioProtoToState, convertStateToRadioProto } from './radio';
-import { createSessionManager, newSession } from './internal/session';
-import { MockClient } from './internal/websocket/mock/websocket';
-import { UIBuilder } from './uibuilder';
-import { Page, PageManager } from './internal/page';
-import { Runtime } from './runtime';
+import { SelectboxState } from '../internal/session/state/selectbox';
+import {
+  convertSelectboxProtoToState,
+  convertStateToSelectboxProto,
+  selectbox,
+} from '../selectbox';
+import { createSessionManager, newSession } from '../internal/session';
+import { MockClient } from '../internal/websocket/mock/websocket';
+import { UIBuilder } from '../uibuilder';
+import { Page, PageManager } from '../internal/page';
+import { Runtime } from '../runtime';
 
-test('convertStateToRadioProto', () => {
+test('convertStateToSelectboxProto', () => {
   const id = uuidv4();
-  const label = 'Test Radio';
+  const label = 'Test Selectbox';
   const value = 1;
   const options = ['Option 1', 'Option 2'];
+  const placeholder = 'Select an option';
   const defaultValue = 0;
   const required = true;
   const disabled = false;
 
-  const state = new RadioState(
+  const state = new SelectboxState(
     id,
     value,
     label,
     options,
+    placeholder,
     defaultValue,
     required,
     disabled,
   );
-  const proto = convertStateToRadioProto(state);
+  const proto = convertStateToSelectboxProto(state);
 
   expect(proto.label).toBe(label);
   expect(proto.value).toBe(value);
   expect(proto.options).toEqual(options);
+  expect(proto.placeholder).toBe(placeholder);
   expect(proto.defaultValue).toBe(defaultValue);
   expect(proto.required).toBe(required);
   expect(proto.disabled).toBe(disabled);
 });
 
-test('convertRadioProtoToState', () => {
+test('convertSelectboxProtoToState', () => {
   const id = uuidv4();
-  const label = 'Test Radio';
+  const label = 'Test Selectbox';
   const value = 1;
   const options = ['Option 1', 'Option 2'];
+  const placeholder = 'Select an option';
   const defaultValue = 0;
   const required = true;
   const disabled = false;
 
-  const tempState = new RadioState(
+  const tempState = new SelectboxState(
     id,
     value,
     label,
     options,
+    placeholder,
     defaultValue,
     required,
     disabled,
   );
-  const proto = convertStateToRadioProto(tempState);
+  const proto = convertStateToSelectboxProto(tempState);
 
-  const state = convertRadioProtoToState(id, proto);
+  const state = convertSelectboxProtoToState(id, proto);
 
   if (!state) {
-    throw new Error('RadioState not found');
+    throw new Error('SelectboxState not found');
   }
 
   expect(state.id).toBe(id);
   expect(state.label).toBe(label);
   expect(state.value).toBe(value);
   expect(state.options).toEqual(options);
+  expect(state.placeholder).toBe(placeholder);
   expect(state.defaultValue).toBe(defaultValue);
   expect(state.required).toBe(required);
   expect(state.disabled).toBe(disabled);
 });
 
-test('radio', () => {
+test('selectbox', () => {
   const sessionId = uuidv4();
   const pageId = uuidv4();
   const session = newSession(sessionId, pageId);
@@ -92,39 +102,40 @@ test('radio', () => {
   const runtime = new Runtime(mockWS, sessionManager, pageManager);
 
   const page = pageManager.getPage(pageId);
-
   if (!page) {
     throw new Error('Page not found');
   }
 
   const builder = new UIBuilder(runtime, session, page);
 
-  const label = 'Test Radio';
+  const label = 'Test Selectbox';
   const initialOptions = ['Option 1', 'Option 2'];
   const options = {
     options: initialOptions,
     defaultValue: 'Option 1',
+    placeholder: 'Select an option',
     required: true,
     disabled: false,
-    formatFunc: (val: string) => val.toUpperCase(),
+    formatFunc: (val: string) => val.toLowerCase(),
   };
 
-  builder.radio(label, options);
+  selectbox(builder, label, options);
 
-  const widgetId = builder.generatePageID('radio', [0]);
-  const state = session.state.getRadio(widgetId);
+  const widgetId = builder.generatePageID('selectbox', [0]);
+  const state = session.state.getSelectbox(widgetId);
 
   const expectedDefaultIndex = initialOptions.indexOf(options.defaultValue);
   const expectedFormattedOptions = initialOptions.map(options.formatFunc);
 
   if (!state) {
-    throw new Error('RadioState not found');
+    throw new Error('SelectboxState not found');
   }
 
   expect(state.id).toBe(widgetId);
   expect(state.label).toBe(label);
   expect(state.value).toBe(expectedDefaultIndex);
   expect(state.options).toEqual(expectedFormattedOptions);
+  expect(state.placeholder).toBe(options.placeholder);
   expect(state.defaultValue).toBe(expectedDefaultIndex);
   expect(state.required).toBe(options.required);
   expect(state.disabled).toBe(options.disabled);
